# swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The LocalAI Rest API.",
        "title": "LocalAI API",
        "contact": {
            "name": "LocalAI",
            "url": "https://localai.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/mudler/LocalAI/master/LICENSE"
        },
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/backends/apply": {
            "post": {
                "tags": [
                    "backends"
                ],
                "summary": "Install backends to LocalAI.",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/localai.GalleryBackend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response",
                        "schema": {
                            "$ref": "#/definitions/schema.BackendResponse"
                        }
                    }
                }
            }
        },
...
        "/backends/delete/{name}": {
            "post": {
                "tags": [
                    "backends"
                ],
                "summary": "delete backends from LocalAI.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backend name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response",
                        "schema": {
                            "$ref": "#/definitions/schema.BackendResponse"
                        }
                    }
                }
            }
        },
...
        "/backends/upgrade/{name}": {
            "post": {
                "tags": [
                    "backends"
                ],
                "summary": "Upgrade a backend",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backend name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response",
                        "schema": {
                            "$ref": "#/definitions/schema.BackendResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "localai.GalleryBackend": {
            "type": "object",
            "properties": {
                "force": {
                    "description": "Force reinstalls the backend even when it is already installed and\nrunnable. Off by default so apply stays idempotent for supervising\napps that ensure their backend on every boot.",
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "schema.BackendResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "status_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}